Add a separate variable for tracking the number of tx queues.
Currently we are using rx_pool->num_queues as the number of tx
queues.  While this is correct, it looks confusing.  Unfortunately
num_queues can't be broken out of the rx_pool into the interface
because the broadcast pool will have to be handled in the future,
which won't have an interface associated with it at all.

Clear up the confusion by having a seperate variable to track the
number of tx queues.