Prevent creation of an invalid nvlist
Increase the maximum nesting depth of an nvlist to 10, and track
the current depth in the nvlist whenever an operationg that adds
or removes an nvlist child of an nvlist.  If an add operation
would violate the maximum depth, the operation fails and the
parent nvlist is set to the error state.

When unpacking an nvlist, ensure that we do not recurse past the
maximum depth limit.  Previously a malicious user could craft a
packed nvlist with an arbitrary level of nesting and the
nvlist_unpack() function could blow its stack when recursing.
This may be a security problem if any current consumers of libnv
call nvlist_unpack() on tainted intput.

If an nvlist is set as a child of another nvlist with
nvlist_move_nvlist then fail the operation and set the parent
nvlist to the error state.