Fix 4BSD race that can cause long scheduling delays
When using the 4BSD scheduler, I was occasionally seeing situations
where a high-priority thread that was bound to a CPU took several
hundred microseconds to be scheduled after it was placed on the
run queue, despite the fact that its CPU was idle at the time.

The root cause is a bug in sched_4bsd's sched_add().  When adding
a bound thread to the run queue of an idle CPU, 4BSD will first
send an AST to the idle CPU via kick_other_cpu() and then add the
thread to the run queue.  When the idle thread receives the AST,
it wakes up and checks for threads in the run queue without taking
any locks (presumably for efficiency).  This opens up a race
window where the idle thread can wake up and check the run queue
before the first thread has managed to add the thread to the run
queue.  When that happens, the idle thread remains running on the
CPU until it is woken again, and on a lightly-loaded system this
can take a long time.

This fix narrows the race window by adding the thread to the
run queue before kicking the idle CPU.  However, I believe that
on architectures with less strict cache coherency than x86 there
may still be a race here.