Refactor network stack state separate from VSI state
Currently the ixl driver has a struct ixl_vsi that it uses to
track both network stack interface state (e.g. the ifnet) and
Virtual Station Interface (VSI) state in the hardware.  However,
when the ixl driver is acting as a PF device for SR-IOV, we will
allocate VSIs for the VFs, but won't have any network stack state
for those VSIs in the ixl driver itself (instead the ixlv driver
will attach to the VF device, which has its own state independent
of the PF driver).

Fix this by introducing a new structure, ixl_ifx, which will be
used to track network stack state.  Refactor both the ixl and
ixlv drivers to move their network stack state into this new
structure, and retain the ixl_vsi structure to track only
information related to details of the VSI allocated from the hw.