Add support for CONFIG_IRQ_MAP VC message
This message is used by the VF to request that its RX and TX
queues be associated with particular MSI-X vectors.  This
configuration is performed via a fairly complicated linked-list
mechanism in the hardware.  The linked list is built up backwards
(from the tail to the head) as it turns out to simplify the logic
significantly.