Add support for ENABLE/DISABLE_QUEUES VC messages
These messages are sent by the VF to enable or disable packet
reception and transmission from its queues.  Despite the fact
that a bitmask of queues is provided in the message, this is
ignored and all queues are either enabled or disabled.  This
matches the behaviour of the Linux PF driver.