Add support for ADD/DEL_ETHER_ADDRESS VC messages
The ADD message is sent by the VF to request that a MAC address
filter be added to the Fortville switch to direct packets destined
to that MAC be routed to the VF's VSI.  The DEL message deletes
a MAC address filter that was previously established.  This has
security implications in a VM environment (a VM could listen in
on packets that it's not entitled to see), so the VF has to have
this capability enabled on it.