Save the correct size of the devinfo in the devinfo
When a VF is created, we have to allocate a new devinfo struct
for its ivar.  The size of this structure is uncertain because
different pci bus subclasses have a different sized structure.
Cache the size of the devinfo in all pci instances devinfo,
so that when we are creating a VF we can easily figure out how
much memory to allocate for the devinfo.