Add some security-related config parameters
Allow the administrator to prevent VFs from sending packets with
a source MAC that is not the VF's MAC (anti-spoof).  Default this
to on.

Allow the administrator to give VFs the ability to override the
MAC address that was specified at creation time.  By default VFs
will not be permitted to do this.

Allow the administrator to give the VFs the ability to enter
promiscuous mode.