Implement PCI SR-IOV method to initialize individual VFs
For each VF that has been allocated, the PF driver must allocate
a Virtual Switch Interface (VSI), which is a virtual port on the
virtual Fortville switch (VEB).

In device_attach() for the PF, we now have to initialize all RX
and TX queues instead just those that will be used by the PF,
as there is no way of know how many will be allocated at runtime
for the VFs.