Fix some errno usages in ixl(4)
There were a couple of places where ixl(4) was returning
an internal error value where an errno was expected.  Fix
these.

In addition, some debugging sysctls would return EAGAIN if
they hit a certain kind of failure.  Unfortunately, if a
sysctl handler returns EAGAIN the kernel takes this literally
and attempts the sysctl again.  This can lead to an infinite
loop if the sysctl handler keeps hitting the same error.  Fix
this by returning a different error code instead.