Refactor ixlv initialization
Initialize an ixlv interface by using a state machine to track
the status of the initialization.  This allows the init to happen
without busy-waiting for the PF driver.  If the VF driver runs
on the host, then busy-waiting can leave the PF driver without
any CPU cycles to perform the VF's request, and eventually the VF
times out and fails.