Fix a regression in RTO(min) for FE TCP
The incast mitigation feature for back-end ethernet inadvertently
decreased the minimum retransmit timeout for front-end connections.
Two bugs were introduced: first, the minimum RTT measurement was
changed from 30 ms to 3 ticks.  This matched the comment above
the code, but this can cause spurious retransmits when
communicating with hosts with a low average RTT but occasional
large latency bursts.

Second, when the units for TCP timeouts were changed from ticks
to sbintime, the rexmit_slop parameter was not updated to match.
This means that the slop value changed to be effectively zero.
The slop is necessary to prevent unnecessary retransmits (and
potentially connection drops) in the face of delayed ACKs from
the remote host.

It is not sufficent to merely restore the functionality of the
200ms slop value, as that would also introduce unacceptable
latency into retransmits on the back-end network.  Instead make
the value configurable on a per-socket basis, with the default
back at the original 200ms.  Update RBM to set the slop to 0,
as additional delay for retransmits on the back-end would just
have a negative impact on recovery from fabric discards due
to incast, and delayed ACKs have already been properly handled.

Bug: https://bugs.west.isilon.com/show_bug.cgi?id=182635

CR: https://reviewboard.west.isilon.com/r/52173/
(cherry picked from commit 84b035ffdc4c351c6adfe2174b40e5c84ba8e084)