Port the callout system to use a heap
The callwheel allows certain operations to be constant-time,
but does not implement a way to find the next callout to run
quickly.  This makes it impossible to accurately schedule timer
interrupts, causing many spurious interrupts, hurting performance.
The problem is most noticeable when callouts are frequently
cancelled or rescheduled for later, as the original timer
interrupt to run that callout will remain scheduled.

Replace the callwheel with a binary heap, which performs most
operations in O(log n) time, but allows for a quick lookup to
determine the next time to schedule an interrupt.