mlx4_en: Move debug counters to per-ring structure
Currently the mlx4_en driver increments a number of port-wide
counters in the rx and tx paths.  This causes the cache line
containing those counters to ping-pong between CPUs when more
than one ring is being used simultaneously, causing a significant
perf hit for frequently-incremented counters.

The driver already has the infrastructure to maintain separate
per-ring counters and then aggregate them together periodically,
so make use of that to fix the cache problems.