Loosen the linkage between a program and a file
When an ebpf_prog is first loaded, it is associated with a
struct file.  However, in principle there is no reason why the
file cannot be destroyed by having all open file descriptors
closed while the program remains active.  Allow this by ensuring
that activations only keep a reference on the program, not the file.