Allow ebpf_prog to share storage with a larger type
Allow ebpf_prog to be allocated as part of a larger object.
struct ebpf_prog must be the first member of the object, but the
any memory after that is fair game for any purpose.  The
reference count in the ebpf_obj subobject will control deallocation
for the whole allocation.