Rework the deallocation path for an activation
When a program deactivation is initiated by the provider side,
quiesce the provider-specific activation state but do not
immediately free the state.  Instead we have to call into the
ebpf module to get it to tear down its end of the activation
after we have released our provider-side lock.

This will allow us to deactivate a program either from the
provider side or the module side with LORs or other races.