Allow a probe activation to be marked persistent
If an activation is marked as "persistent", then the program will
remain attached to the probe until something detaches us from
the probe side (e.g. for a syscall probe, the process being traced
exits).

For a normal activation, if all references to the program's file
descriptor are closed, then the program will be detached
immediately.